split AppsScreen into controller and components
