+++ BrightID/src/components/Apps/AppsScreen.tsx
@@ -1,92 +1,61 @@
-import React, {
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-  useRef,
-} from 'react';
+import React, { useRef } from 'react';
 import {
-  Alert,
   StyleSheet,
   View,
   Text,
-  TextInput,
   RefreshControl,
   Animated,
-  TouchableOpacity,
   Platform,
   Easing,
 } from 'react-native';
 import { useTranslation } from 'react-i18next';
-import { any, propEq, find } from 'ramda';
-import {
-  useFocusEffect,
-  useNavigation,
-  useRoute,
-} from '@react-navigation/native';
 import { useHeaderHeight } from '@react-navigation/stack';
-import _ from 'lodash';
 import Spinner from 'react-native-spinkit';
-import Svg, { Path } from 'react-native-svg';
-import {
-  linkedContextTotal,
-  selectAllLinkedContexts,
-  selectAllLinkedSigs,
-} from '@/reducer/appsSlice';
-import { useDispatch, useSelector } from '@/store';
+import { useFocusEffect } from '@react-navigation/native';
 import EmptyList from '@/components/Helpers/EmptyList';
 import { ORANGE, BLUE, WHITE } from '@/theme/colors';
 import { DEVICE_LARGE } from '@/utils/deviceConstants';
-import {
-  fetchApps,
-  selectAllApps,
-  selectPendingLinkedContext,
-} from '@/actions';
 import { fontSize } from '@/theme/fonts';
-import { NodeApiContext } from '@/components/NodeApiGate';
-import { isVerified } from '@/utils/verifications';
 import AppCard from './AppCard';
-import { handleV5App, handleV6App } from './model';
-import { AppsRoute } from '@/components/Apps/types';
 import AnimatedLinearGradient from './AnimatedLinearGradient';
+import AppsScreenFilter from '@/components/Apps/AppsScreenFilter';
 
-export const AppsScreen = () => {
-  const filters = [
-    { name: 'All Apps', id: 1 },
-    { name: 'Linked', id: 2 },
-    { name: 'Verified', id: 3 },
-  ];
+type Props = {
+  sponsoringApp: AppInfo | undefined;
+  pendingLink: ContextInfo | undefined;
+  isSponsored: boolean;
+  totalApps: number;
+  linkedContextsCount: number;
+  totalVerifiedApps: number;
+  activeFilter: number;
+  searchTerm: string;
+  setFilter: (filter: number) => void;
+  setSearch: (term: string) => void;
+  filteredApps: AppInfo[];
+  refreshing: boolean;
+  refreshApps: () => void;
+};
 
-  const dispatch = useDispatch();
-  const navigation = useNavigation();
-  const route = useRoute<AppsRoute>();
-  const api = useContext(NodeApiContext);
+export const AppsScreen = ({
+  sponsoringApp,
+  pendingLink,
+  isSponsored,
+  totalApps,
+  linkedContextsCount,
+  totalVerifiedApps,
+  activeFilter,
+  setFilter,
+  searchTerm,
+  setSearch,
+  filteredApps,
+  refreshing,
+  refreshApps,
+}: Props) => {
   const headerHeight = useHeaderHeight();
   const { t } = useTranslation();
+  useFocusEffect(refreshApps);
 
-  const apps = useSelector(selectAllApps);
-  const isSponsored = useSelector(
-    (state: State) => state.user.isSponsored || state.user.isSponsoredv6,
-  );
-  const linkedContext = useSelector(selectAllLinkedContexts);
-  const linkedContextsCount = useSelector(linkedContextTotal);
-  const selectLinkedSigs = useSelector(selectAllLinkedSigs);
-  const pendingLink = useSelector(selectPendingLinkedContext);
-  const userVerifications = useSelector(
-    (state: State) => state.user.verifications,
-  );
-
-  const [refreshing, setRefreshing] = useState(false);
-  const [sponsoringApp, setSponsoringApp] = useState<AppInfo | undefined>(
-    undefined,
-  );
-  const [totalApps, setTotalApps] = useState(0);
-  const [totalVerifiedApp, setTotalVerifiedApp] = useState(0);
-
-  // filter state
-  const [search, setSearch] = useState('');
-  const [activefilter, SetFilter] = useState(0);
-  const [filteredApp, setFilteredApp] = useState(apps);
+  const scrollViewRef = useRef(null);
 
   // Animation
   const scrollY = useRef(new Animated.Value(0)).current;
@@ -134,133 +103,6 @@ export const AppsScreen = () => {
     easing: Easing.linear,
     useNativeDriver: false, // <-- neccessary
   });
-
-  const refreshApps = useCallback(() => {
-    setRefreshing(true);
-    dispatch(fetchApps(api))
-      .then(() => {
-        setRefreshing(false);
-      })
-      .catch((err) => {
-        console.log(err.message);
-        setRefreshing(false);
-      });
-  }, [api, dispatch]);
-
-  const handleV5DeepLink = useCallback(() => {
-    const context = route.params?.context;
-    const isValidApp = any(propEq('id', context))(apps);
-    const isValidContext = any(propEq('context', context))(apps);
-    // legacy apps send context in the deep link but soulbound apps send app
-    if (isValidApp || isValidContext) {
-      handleV5App(route.params, setSponsoringApp, api);
-    } else {
-      Alert.alert(
-        t('apps.alert.title.invalidContext'),
-        t('apps.alert.text.invalidContext', { context: `${context}` }),
-      );
-    }
-    // reset params
-    navigation.setParams({
-      baseUrl: '',
-      context: '',
-      contextId: '',
-    });
-  }, [navigation, route.params, apps, api, t]);
-
-  const handleV6DeepLink = useCallback(() => {
-    const appId = route.params?.context;
-    const appInfo = find(propEq('id', appId))(apps) as AppInfo;
-    if (api && appInfo && appInfo.usingBlindSig) {
-      handleV6App(route.params, setSponsoringApp, api);
-    } else {
-      Alert.alert(
-        t('apps.alert.title.invalidApp'),
-        t('apps.alert.text.invalidApp', { app: `${appId}` }),
-      );
-    }
-    // reset params
-    navigation.setParams({
-      context: '',
-      contextId: '',
-    });
-  }, [route.params, apps, navigation, api, t]);
-
-  useEffect(() => {
-    if (apps.length > 0 && route.params?.baseUrl) {
-      handleV5DeepLink();
-    } else if (apps.length > 0 && route.params?.context) {
-      handleV6DeepLink();
-    }
-  }, [apps, handleV5DeepLink, handleV6DeepLink, route.params]);
-
-  // handle filter
-  useEffect(() => {
-    const allApps = apps.filter((app) => {
-      const isLinked = linkedContext.find(
-        (link) => link.context === app.context,
-      );
-      const isLinkedTemp = isLinked && isLinked.state === 'applied';
-      return !app.testing || isLinkedTemp;
-    });
-
-    // filter linked app
-    const linkedApps = allApps.filter((app) => {
-      const appSig = selectLinkedSigs.filter((sig) => sig.app === app.id);
-      const isLinked = linkedContext.find(
-        (link) => link.context === app.context,
-      );
-      const isLinkedTemp = isLinked && isLinked.state === 'applied';
-      return (app.usingBlindSig && appSig.length > 0) || isLinkedTemp;
-    });
-
-    // filter verified app
-    const verifiedApps = allApps.filter((app) => {
-      let count = 0;
-      for (const verification of app.verifications) {
-        const verified = isVerified(
-          _.keyBy(userVerifications, (v) => v.name),
-          verification,
-        );
-        if (verified) {
-          count++;
-        }
-      }
-
-      return count === app.verifications.length;
-    });
-
-    setTotalApps(allApps.length);
-    setTotalVerifiedApp(verifiedApps.length);
-
-    let filteredApp: AppInfo[] = [];
-    if (activefilter === 0) {
-      filteredApp = allApps;
-    } else if (activefilter === 1) {
-      filteredApp = linkedApps;
-    } else if (activefilter === 2) {
-      filteredApp = verifiedApps;
-    }
-
-    // filter using search bar
-    if (search !== '') {
-      const filterResult = filteredApp.filter(
-        (app) => app.name.toLowerCase().indexOf(search.toLowerCase()) !== -1,
-      );
-      setFilteredApp(filterResult);
-    } else {
-      setFilteredApp(filteredApp);
-    }
-  }, [
-    search,
-    activefilter,
-    apps,
-    selectLinkedSigs,
-    linkedContext,
-    userVerifications,
-  ]);
-
-  useFocusEffect(refreshApps);
 
   const OverallDescription = () => {
     let msg: string, waiting: boolean;
@@ -333,67 +175,14 @@ export const AppsScreen = () => {
             You meet the verification criteria of
           </Text>
           <Text style={[styles.detail, { textAlign: 'right' }]}>
-            {totalVerifiedApp} <Text style={styles.detailLabel}>apps</Text>
+            {totalVerifiedApps} <Text style={styles.detailLabel}>apps</Text>
           </Text>
         </View>
       </AnimatedLinearGradient>
     );
   };
 
-  const Filter = () => {
-    return (
-      <>
-        <Animated.View
-          testID="searchBarBackground"
-          style={[
-            styles.searchBackground,
-            { top: headerHeight, opacity: fadeBackgroundSearch },
-          ]}
-        />
-
-        <Animated.View
-          style={[
-            styles.searchContainer,
-            {
-              top: headerHeight + 10,
-              transform: [{ translateY: translateYSearch }],
-            },
-          ]}
-        >
-          <TextInput
-            style={[styles.shadow, styles.textInput]}
-            onChangeText={(value) => setSearch(value)}
-            placeholder="App name"
-            value={search}
-          />
-
-          <View style={styles.filterContainer}>
-            {filters.map((item, index) => (
-              <TouchableOpacity
-                key={index}
-                style={[
-                  styles.filterItemContainer,
-                  { backgroundColor: index === activefilter ? ORANGE : WHITE },
-                ]}
-                onPress={() => SetFilter(index)}
-              >
-                <Text
-                  style={{
-                    fontFamily: 'Poppins-Medium',
-                    fontSize: 12,
-                    color: index === activefilter ? WHITE : ORANGE,
-                  }}
-                >
-                  {item.name}
-                </Text>
-              </TouchableOpacity>
-            ))}
-          </View>
-        </Animated.View>
-      </>
-    );
-  };
-
+  console.log(`Rendering appsscreen`);
   return (
     <>
       <View style={styles.container} testID="appsScreen">
@@ -407,11 +196,19 @@ export const AppsScreen = () => {
 
         <OverallDescription />
 
-        <Filter />
+        <AppsScreenFilter
+          filter={activeFilter}
+          searchTerm={searchTerm}
+          setFilter={setFilter}
+          setSearchTerm={setSearch}
+          fadeBackgroundSearch={fadeBackgroundSearch}
+          translateYSearch={translateYSearch}
+        />
 
         <Animated.FlatList
+          ref={(ref) => (scrollViewRef.current = ref)}
           testID="appsList"
-          data={filteredApp}
+          data={filteredApps}
           onScroll={handleScroll}
           scrollEventThrottle={16}
           numColumns={2}
@@ -438,12 +235,6 @@ export const AppsScreen = () => {
 };
 
 const styles = StyleSheet.create({
-  orangeTop: {
-    backgroundColor: ORANGE,
-    height: DEVICE_LARGE ? 70 : 65,
-    width: '100%',
-    zIndex: 1,
-  },
   container: {
     flex: 1,
   },
@@ -525,13 +316,6 @@ const styles = StyleSheet.create({
     fontSize: 20,
     color: WHITE,
   },
-  textInput: {
-    backgroundColor: 'white',
-    width: '90%',
-    height: 50,
-    borderRadius: 5,
-    paddingHorizontal: 10,
-  },
   wrapperColumn: {
     justifyContent: 'space-between',
     marginBottom: 20,
@@ -543,37 +327,6 @@ const styles = StyleSheet.create({
     backgroundColor: 'white',
     borderTopLeftRadius: 10,
     borderTopRightRadius: 10,
-  },
-  filterContainer: {
-    width: '90%',
-    flexDirection: 'row',
-    marginTop: 10,
-  },
-  filterItemContainer: {
-    borderWidth: 1,
-    borderColor: ORANGE,
-    padding: 5,
-    borderRadius: 5,
-    marginRight: 5,
-  },
-  searchContainer: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    alignSelf: 'center',
-    alignItems: 'center',
-    width: '100%',
-    height: 100,
-    zIndex: 6,
-    // marginTop: 40,
-    overflow: 'visible',
-  },
-  searchBackground: {
-    position: 'absolute',
-    width: '100%',
-    height: 120,
-    backgroundColor: `white`,
-    zIndex: 5,
   },
 });
